+++ packages/faas-plugin-common/src/cli.ts
@@ -1,0 +1,111 @@
+import { CommandHookCore } from '@midwayjs/command-core';
+import * as minimist from 'minimist';
+import { join } from 'path';
+import { loadSpec } from './utils/loadSpec';
+import { PluginManager as CommandPlugin } from './plugins/pluginManager';
+import { Invoke as CommandInvoke } from './plugins/invoke';
+import { Test as CommandTest } from './plugins/test';
+
+export * from './plugins/invoke/main';
+export const InvokeClass = CommandInvoke;
+export const Test = CommandTest;
+
+export class CLI {
+  argv: any;
+  providerName: string;
+  core: any;
+  spec: any;
+  commands: string[];
+  cwd;
+
+  constructor(argv) {
+    this.argv = minimist(argv.slice(2));
+    this.commands = [].concat(this.argv._);
+    this.loadSpec();
+    this.providerName = (this.spec.provider && this.spec.provider.name) || '';
+    this.cwd = process.cwd();
+    this.core = new CommandHookCore({
+      config: {
+        servicePath: this.cwd,
+      },
+      commands: this.commands,
+      service: this.spec,
+      provider: this.providerName,
+      options: this.argv,
+      log: this.loadLog(),
+      extensions: this.loadExtensions(),
+    });
+    this.loadDefaultPlugin();
+    this.loadPlatformPlugin();
+    this.loadUserPlugin();
+  }
+
+  // 加载默认插件
+  loadDefaultPlugin() {
+    this.loadCommandPlugin();
+    this.loadCommandInvoke();
+    this.loadCommandTest();
+  }
+
+  loadCommandPlugin() {
+    this.core.addPlugin(CommandPlugin);
+  }
+
+  loadCommandInvoke() {
+    this.core.addPlugin(CommandInvoke);
+  }
+
+  loadCommandTest() {
+    this.core.addPlugin(CommandTest);
+  }
+
+  // 加载平台方插件
+  loadPlatformPlugin() {}
+
+  // 加载cli拓展
+  loadExtensions() {
+    return {};
+  }
+
+  loadSpec() {
+    this.spec = loadSpec(this.cwd);
+  }
+
+  // 加载命令行输出及报错
+  loadLog() {
+    return console;
+  }
+
+  error(errMsg) {
+    console.log('errMsg', errMsg);
+    process.exit(1);
+  }
+
+  loadUserPlugin() {
+    if (!this.spec || !this.spec.plugins) {
+      return;
+    }
+    for (const plugin of this.spec.plugins) {
+      if (/^npm:/.test(plugin) || /^local:/.test(plugin)) {
+        this.core.addPlugin(plugin);
+      } else if (/^\./.test(plugin)) {
+        this.core.addPlugin(`local:${this.providerName}:${plugin}`);
+      }
+    }
+  }
+
+  loadRelativePlugin(dirPath, path) {
+    try {
+      const localPlugin = require(join(this.cwd, dirPath, path));
+      this.core.addPlugin(localPlugin);
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+
+  async start() {
+    await this.core.ready();
+    await this.core.invoke();
+  }
+}
