+++ packages/faas-plugin-common/src/index.ts
@@ -1,109 +1,3 @@
-const minimist = require('minimist');
-const CoreClass = require('@midwayjs/command-core');
-const { join } = require('path');
-import { loadSpec } from './utils/loadSpec';
-import CommandPlugin from './plugins/pluginManager';
-import CommandInvoke from './plugins/invoke';
-import CommandTest from './plugins/test';
-
-const baseDir = process.cwd();
-export * from './plugins/invoke/main';
-export const InvokeClass = CommandInvoke;
-export const Test = CommandTest;
-export class Cli {
-  argv: any;
-  providerName: string;
-  core: any;
-  spec: any;
-  commands: string[];
-  constructor(argv) {
-    this.argv = minimist(argv.slice(2));
-    this.commands = [].concat(this.argv._);
-    this.loadSpec();
-    this.providerName = (this.spec.provider && this.spec.provider.name) || '';
-    this.core = new CoreClass({
-      config: {
-        servicePath: baseDir,
-      },
-      commands: this.commands,
-      service: this.spec,
-      provider: this.providerName,
-      options: this.argv,
-      log: this.loadLog(),
-      extensions: this.loadExtensions()
-    });
-    this.loadDefaultPlugin();
-    this.loadPlatformPlugin();
-    this.loadUserPlugin();
-  }
-
-  // 加载默认插件
-  loadDefaultPlugin() {
-    this.loadCommandPlugin();
-    this.loadCommandInvoke();
-    this.loadCommandTest();
-  }
-
-  loadCommandPlugin() {
-    this.core.addPlugin(CommandPlugin);
-  }
-
-  loadCommandInvoke() {
-    this.core.addPlugin(CommandInvoke);
-  }
-
-  loadCommandTest() {
-    this.core.addPlugin(CommandTest);
-  }
-
-  // 加载平台方插件
-  loadPlatformPlugin() {
-  }
-
-  // 加载cli拓展
-  loadExtensions() {
-    return {};
-  }
-
-  loadSpec() {
-    this.spec = loadSpec(baseDir);
-  }
-
-  // 加载命令行输出及报错
-  loadLog() {
-    return console;
-  }
-
-  error(errMsg) {
-    console.log('errMsg', errMsg);
-    process.exit(1);
-  }
-
-  loadUserPlugin() {
-    if (!this.spec || !this.spec.plugins) {
-      return;
-    }
-    for (const plugin of this.spec.plugins) {
-      if (/^npm:/.test(plugin) || /^local:/.test(plugin)) {
-        this.core.addPlugin(plugin);
-      } else if (/^\./.test(plugin)) {
-        this.core.addPlugin(`local:${this.providerName}:${plugin}`);
-      }
-    }
-  }
-
-  loadRelativePlugin(dirPath, path) {
-    try {
-      const localPlugin = require(join(baseDir, dirPath, path));
-      this.core.addPlugin(localPlugin);
-      return true;
-    } catch (e) {
-      return false;
-    }
-  }
-
-  async start() {
-    await this.core.ready();
-    await this.core.invoke();
-  }
-}
+export * from './cli';
+export { Invoke as InvokeClass } from './plugins/invoke';
+export { Test } from './plugins/test';
